requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
pygithub>=2.2.0
sentence-transformers>=2.5.0
chromadb>=0.4.22
openai>=1.12.0
#pdfkit>=1.0.0
markdown>=3.5.2
python-multipart>=0.0.9
pytest>=7.4.0
